oaback/requirements.txt
aes==1.2.0
amqp==5.2.0
ansicon==1.89.0
arrow==1.3.0
asgiref==3.7.2
attrs==23.2.0
billiard==4.2.0
blessed==1.20.0
celery==5.4.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==42.0.5
Django==5.0.3
django-cors-headers==4.3.1
django-environ==0.11.2
django-filter==24.2
django-picklefield==3.1
django-redis==5.4.0
django-shortuuidfield==0.1.3
djangorestframework==3.15.1
dnspython==2.6.1
drf-spectacular==0.27.2
drf-spectacular-sidecar==2024.4.1
environ==1.0
et-xmlfile==1.1.0
eventlet==0.35.2
Faker==24.2.0
gevent==24.2.1
greenlet==3.0.3
idna==3.6
inflection==0.5.1
jinxed==1.2.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kombu==5.3.5
mysqlclient==2.1.1
Naked==0.1.32
numpy==1.26.4
openpyxl==3.1.2
pandas==2.2.1
pillow==10.2.0
prompt-toolkit==3.0.43
pycparser==2.21
pycryptodome==3.20.0
pycryptodomex==3.20.0
PyJWT==2.8.0
PyMySQL==1.1.0
pyOpenSSL==24.0.0
pypinyin==0.51.0
python-alipay-sdk==3.3.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
redis==5.0.3
referencing==0.34.0
requests==2.31.0
ronglian-sms-sdk==1.0.0
rpds-py==0.18.0
setuptools==69.2.0
shellescape==3.8.1
shortuuid==1.0.13
six==1.16.0
sqlparse==0.4.4
types-python-dateutil==2.9.0.20240316
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
xlrd==2.0.1
zope.event==5.0
zope.interface==6.2
djangorestframework-simplejwt